refactor(fsd): Step 5 — create widgets/ layer (app-shell + voc-workspace)
Move components/layout/* → widgets/app-shell/ and VocListPage → widgets/voc-workspace/.
Update router.tsx + pages/VocPage.tsx to use @widgets/* alias.
VocListPage internal imports updated from relative to @features/*.
All 374 tests pass, typecheck clean.
