+++ frontend/src/pages/VocPage.tsx
@@ -1,4 +1,4 @@
-import { VocListPage } from '../components/voc/VocListPage';
+import { VocListPage } from '@widgets/voc-workspace';
 
 export default function VocPage() {
   return <VocListPage />;
