+++ frontend/src/router.tsx
@@ -1,10 +1,9 @@
 import { lazy, Suspense } from 'react';
 import { env } from '@shared/config/env';
 import { createBrowserRouter, Navigate } from 'react-router-dom';
-import { AppShell } from './components/layout/AppShell';
+import { AppShell, PageTitle } from '@widgets/app-shell';
 import { EmptyState } from '@shared/ui/empty-state';
 import { LoadingState } from '@shared/ui/skeleton';
-import { PageTitle } from './components/layout/PageTitle';
 
 const MockLoginPage = env.AUTH_MODE === 'mock' ? lazy(() => import('./pages/MockLoginPage')) : null;
