+++ frontend/src/widgets/app-shell/Sidebar.tsx
