+++ frontend/src/widgets/app-shell/TopBar.tsx
