+++ frontend/src/widgets/app-shell/index.ts
@@ -1,0 +1,4 @@
+export { AppShell } from './AppShell';
+export { PageTitle } from './PageTitle';
+export { Sidebar } from './Sidebar';
+export { TopBar } from './TopBar';
