+++ frontend/src/widgets/voc-workspace/__tests__/VocListPage.integration.test.tsx
