+++ frontend/src/widgets/voc-workspace/index.ts
@@ -1,0 +1,1 @@
+export { VocListPage } from './VocListPage';
